refactor(macros): place container/field/variant in separate files under `meta`
